+++ modules/Render/RenderVulkan/src/Resources/VkTexture.h
@@ -29,7 +29,7 @@ bool dsVkTexture_copyData(dsResourceManager* resourceManager, dsCommandBuffer* c
 	uint32_t layers, const void* data, size_t size);
 bool dsVkTexture_copy(dsResourceManager* resourceManager, dsCommandBuffer* commandBuffer,
 	dsTexture* srcTexture, dsTexture* dstTexture, const dsTextureCopyRegion* regions,
-	size_t regionCount);
+	uint32_t regionCount);
 bool dsVkTexture_generateMipmaps(dsResourceManager* resourceManager, dsCommandBuffer* commandBuffer,
 	dsTexture* texture);
 bool dsVkTexture_getData(void* result, size_t size, dsResourceManager* resourceManager,
